Further performance improvements
* Don't sort creeps in jumper tower, instead find closest via linear scan.
* Use HashSet in BeamTower and WaveTower, as in very late game there can be
200+ creeps, and the wave/beam may have hit all of them. The HashSet is more
efficient at that stage.
* Expand use of CreepGrid to most towers, adding new filter methods.
* Optimisations when a bullet crosses multiple cells in CreepGrid.
* Minor optimisation to CreepGrid.excluding().
* Fix issue where ghost cost was only using an int.
* Improve calculation of when a Wave is off screen.
* Use Math.atan2 instead of custom code in Vector2D.angle()

Also added an ant target to run unit tests.

+++ logic/Formulae.java
@@ -72,8 +72,8 @@ public class Formulae {
     return (long) (Math.pow(2, currentLevel - 1)) * 250;
   }
 
-  public static int towerCost(int numTowers, int numOfThisType) {
-    return (int) (Math.pow(1.05, numTowers) * Math.pow(1.1, numOfThisType) * 1000);
+  public static long towerCost(int numTowers, int numOfThisType) {
+    return (long) (Math.pow(1.05, numTowers) * Math.pow(1.1, numOfThisType) * 1000);
   }
 
   public static long sellValue(Tower t, int numTowers, int numOfThisType) {
